fix getttywidth on BSD systems
TIOCGSIZE has only been left on BSD system for compatibility purpose.
DragonFlyBSD, FreeBSD, NetBSD and OpenBSD all define TIOCGWINSZ for
quite a long time now. It is therefore a bad assumption to assume that
if TIOCGSIZE is defined, we are on an old system and that it shall be
used.