WIP text export refactoring
Major refactoring of the way that width for columns is computed. It
includes removal of some function parameters that have become obsolete
and addition of new functions to compute maxwidths for columns in text
export. This is still a work in progress but as there is already quite
a few changes, committing for now.