avoid stat'ing remote fs when local flag is given
This prevents dfc from hanging on remote file system.

Note that this was an issue on linux only since for the BSD
implementation, the MNT_NOWAIT flag is given to getmntinfo(3) which
prevents the underlying getfsstat(2) call from blocking when trying to
get information from a file system that is unable to respond.