Show every filesystem only once, unless -a is given.
They were listed multiple times when bind mounts or btrfs/zfs subvolumes
are in use.  And in those cases, a single filesystem is often mounted many,
many times.

Filesystems whose backing store is not an absolute path are assumed to be
unique.  This is not always correct (they can be bind-mounted).

In case of duplicates, the mount with a shorter path is considered more
important.