Hide /run mounts unless they're more than half full, unless -a is given.
On modern Debian-based systems there's usually 3 of them, on Fedora 23 4.

I'm not sure what's the good place to put this code in, as you prefer to
separate logic from presentation.  It doesn't fit into is_mnt_ignore() as
it's used by some distributions at least on linux, freebsd and hurd and
that function is platform-specific.