fix a bug which messes up colors in HTML export when config file is used
This is an old (at least 4 years old) bug that happened only if `dfc`
exports its result to HTML AND a configuration file is used. In this
case, the HTML output color variables would be set to gibberish due to
bad memory handling. Apparently, exporting to HTML while also having a
configuration file is not a popular scenario.

This commit fixes the issue while it also cleans things up a bit by
making the corresponding fields in the global config structure char
arrays instead of const char pointers.

Note that this commit also improves error messages that are output when
a wrong value is given to a configuration file key by also telling which
key has a wrong value.