improved the summary bar.
in absolute graph mode, it is wider than the other bars, so I needed
support for flexible graph / bar width. before this commit, the color
steps in the absolute summary graph would be on the same level as the
ones above but not at the correct percentages, think something like this:
sda1 ggggggyyyyrr--
sda2 ggggggy-------
sda3 ggggggyy------
sum: ggggggyyyyrrrrrrrrrrrrrrrrr---------------
but it should look like this:
     ggggggggggggggggggyyyyyyyrr---------------
This issue is fixed in this commit.