Fixed a bug where the details of the total bar would be off.
If dfc decides at runtime to only display short  bars instead of wide
bars, because the tty is too narrow, the `max.bar` still indicates
`GRAPHBAR_WIDE` but it should be `GRAPHBAR_SHORT`.
So for calculating the detail fields offset I am now using the ternary
thing instead of `max.bar`.

This requires another look into it because currently, the titlebar
is off if we are using the `-w` option and the tty is too narrow.